Fix header levels
We were inserting the article's title as a level 1 header inside the
articles, and then having all sections on further levels. That creates
problems with tex rendering (TODO) and TOC generation.

We now add the document title as a <h1 id="title> in the html
template, and leave the document sections with proper header levels.